+++ Automate/release-notes.md
@@ -3,6 +3,7 @@
 # Release notes
 ## Upcoming release
 * Extended Automate API to simplify custom machines.
+* Fixed error when the game has invalid fish data.
 * Fixed tree tappers having the wrong automation position in Automate 1.11.
 * Fixed gates not working as connectors.