+++ TractorMod/release-notes.md
@@ -1,4 +1,7 @@
 # Release notes
+## Upcoming release
+* Fixed stable no longer in Robin's carpenter menu after building a tractor garage.
+
 ## 4.7.2
 * Fixed horses becoming tractors in some cases. (If you saved after that happened, [see this forum comment](https://community.playstarbound.com/threads/tractor-mod.136649/page-14#post-3319770)).
 * Fixed tractors sometimes not in their garage after loading a save.